requirements.txt
django
django-mptt
pytils
